refactor: split Bits god-context into Balance/Debit/Transaction sub-contexts
- Bits.Balance: balance CRUD with Nebulex cache decorators
- Bits.Debit: debit CRUD + activate_translations_for
- Bits.Transaction: transaction CRUD + process_bits_transaction
- Fix: broadcasts moved outside Ecto.Multi steps to prevent
  phantom updates on transaction rollback
- Bits facade: defdelegate to sub-contexts for backwards compat
- @cache_ttl as module attribute (no magic numbers)

Closes #287
