+++ lib/stream_closed_captioner_phoenix/bits.ex
@@ -1,608 +1,44 @@
 defmodule StreamClosedCaptionerPhoenix.Bits do
-  @moduledoc """
-  The Bits context.
-  """
-
-  use Nebulex.Caching
-
-  import Ecto.Query, warn: false
-
-  alias StreamClosedCaptionerPhoenix.Accounts.User
-  alias StreamClosedCaptionerPhoenix.Bits.BitsBalanceDebit
-  alias StreamClosedCaptionerPhoenix.Bits.BitsBalanceDebitQueries
-  alias StreamClosedCaptionerPhoenix.Bits.BitsBalanceQueries
-  alias StreamClosedCaptionerPhoenix.Bits.BitsTransactionQueries
-  alias StreamClosedCaptionerPhoenix.AuditLog
-  alias StreamClosedCaptionerPhoenix.Cache
-  alias StreamClosedCaptionerPhoenix.Repo
-
-  @cache_ttl :timer.minutes(2)
-
-  def bits_transactions_and_debits_for_user(user_id, offset, limit) do
-    records =
-      BitsTransactionQueries.get_bits_transactions_and_debits_for_user(user_id)
-      |> limit(^limit)
-      |> offset(^offset)
-      |> Repo.all()
-
-    total_records =
-      Repo.count(BitsTransactionQueries.get_bits_transactions_and_debits_for_user(user_id))
-
-    %{records: records, total_records: total_records}
-  end
-
-  def activate_translations_for(%User{} = user) do
-    result =
-      Ecto.Multi.new()
-      |> Ecto.Multi.run(:bits_balance_check, bits_balance_check(user))
-      |> Ecto.Multi.run(:debit, fn _repo, _ ->
-        create_bits_balance_debit(user, %{amount: 500})
-      end)
-      |> Ecto.Multi.run(:update_balance, &update_bits_balance_transaction/2)
-      |> Ecto.Multi.run(:broadcast, broadcast_updated_bits_balance(user))
-      |> Repo.transaction()
-
-    case result do
-      {:ok, %{debit: debit, update_balance: updated_balance}} ->
-        AuditLog.info("bits.translation_activated", %{
-          user_id: user.id,
-          debit_id: debit.id,
-          amount: debit.amount,
-          balance: updated_balance.balance
-        })
-
-      {:error, :bits_balance_check, :insufficent_balance, _} ->
-        AuditLog.warn("bits.translation_activation_failed", %{
-          user_id: user.id,
-          reason: :insufficent_balance
-        })
-
-      {:error, step, reason, _} ->
-        AuditLog.warn("bits.translation_activation_failed", %{
-          user_id: user.id,
-          step: step,
-          reason: audit_failure_reason(reason)
-        })
-    end
-
-    result
-  end
-
-  defp bits_balance_check(user) do
-    fn repo, _ ->
-      bits_balance =
-        BitsBalanceQueries.with_user_id(user.id)
-        |> limit(1)
-        |> repo.one(lock: "FOR UPDATE")
-
-      if bits_balance && bits_balance.balance >= 500 do
-        {:ok, bits_balance}
-      else
-        {:error, :insufficent_balance}
-      end
-    end
-  end
-
-  defp update_bits_balance_transaction(_repo, %{bits_balance_check: bits_balance, debit: debit}) do
-    new_balance = bits_balance.balance - debit.amount
-    update_bits_balance(bits_balance, %{balance: new_balance})
-  end
-
-  defp broadcast_updated_bits_balance(user) do
-    fn _repo, %{update_balance: %{balance: balance}} ->
-      StreamClosedCaptionerPhoenixWeb.Endpoint.broadcast(
-        "captions:#{user.id}",
-        "translationActivated",
-        %{enabled: true, balance: balance}
-      )
-
-      {:ok, true}
-    end
-  end
-
-  @doc """
-  Returns the list of bits_balance_debits.
-
-  ## Examples
-
-      iex> list_bits_balance_debits()
-      [%BitsBalanceDebit{}, ...]
-
-  """
-  def list_bits_balance_debits do
-    BitsBalanceDebitQueries.all()
-    |> Repo.all()
-  end
-
-  @doc """
-  Returns the list of bits_balance_debits by user.
-
-  ## Examples
-
-      iex> list_users_bits_balance_debits()
-      [%BitsBalanceDebit{}, ...]
-
-  """
-  def list_users_bits_balance_debits(%{id: id}) do
-    BitsBalanceDebitQueries.with_user_id(id)
-    |> Repo.all()
-  end
-
-  @doc """
-  Gets a single bits_balance_debit.
-
-  Raises `Ecto.NoResultsError` if the Bits balance debit does not exist.
-
-  ## Examples
-
-      iex> get_bits_balance_debit!(123)
-      %BitsBalanceDebit{}
-
-      iex> get_bits_balance_debit!(456)
-      ** (Ecto.NoResultsError)
-
-  """
-  def get_bits_balance_debit!(id), do: BitsBalanceDebitQueries.with_id(id) |> Repo.one!()
-
-  def get_users_bits_balance_debit!(%{id: user_id}, id) do
-    BitsBalanceDebitQueries.with_user_id(user_id)
-    |> BitsBalanceDebitQueries.with_id(id)
-    |> limit(1)
-    |> Repo.one!()
-  end
-
-  @doc """
-  Gets a get_user_active_debit for the user_id that has occurred in the past 24 hours.
-
-  ## Examples
-
-      iex> get_bits_balance_debit!(123)
-      %BitsBalanceDebit{}
-
-      iex> get_bits_balance_debit!(456)
-      nil
-
-  """
-  def get_user_active_debit(user_id) do
-    BitsBalanceDebitQueries.with_user_id(user_id)
-    |> BitsBalanceDebitQueries.less_than_one_day_ago()
-    |> limit(1)
-    |> Repo.one()
-  end
-
-  @doc """
-  Returns a consistent `{balance, debit}` snapshot read atomically from the
-  database, bypassing the cache. Use this when balance and activation state must
-  reflect the same point-in-time (e.g. GraphQL resolvers).
-  """
-  def get_translation_snapshot(user_id) do
-    debit = get_user_active_debit(user_id)
-    balance = BitsBalanceQueries.with_user_id(user_id) |> limit(1) |> Repo.one()
-    {balance, debit}
-  end
-
-  @decorate cacheable(
-              cache: Cache,
-              key: {BitsBalanceDebit, user_id},
-              opts: [ttl: @cache_ttl]
-            )
-  def user_active_debit_exists?(user_id) do
-    BitsBalanceDebitQueries.with_user_id(user_id)
-    |> BitsBalanceDebitQueries.less_than_one_day_ago()
-    |> Repo.exists?()
-  end
-
-  @doc """
-  Creates a bits_balance_debit.
-
-  ## Examples
-
-      iex> create_bits_balance_debit(user, %{field: value})
-      {:ok, %BitsBalanceDebit{}}
-
-      iex> create_bits_balance_debit(user, %{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  @decorate cache_evict(cache: Cache, key: {BitsBalanceDebit, user.id})
-  @decorate cache_evict(cache: Cache, key: {BitsBalance, user.id})
-  def create_bits_balance_debit(user, attrs \\ %{}) do
-    result =
-      user
-      |> Ecto.build_assoc(:bits_balance_debits)
-      |> BitsBalanceDebit.changeset(attrs)
-      |> Repo.insert()
-
-    case result do
-      {:ok, debit} ->
-        AuditLog.info("bits.debit_created", %{
-          user_id: user.id,
-          debit_id: debit.id,
-          amount: debit.amount
-        })
-
-      {:error, _changeset} ->
-        AuditLog.warn("bits.debit_create_failed", %{user_id: user.id})
-    end
-
-    result
-  end
-
-  alias StreamClosedCaptionerPhoenix.Bits.BitsBalance
-
-  @doc """
-  Returns the list of bits_balances.
-
-  ## Examples
-
-      iex> list_bits_balances()
-      [%BitsBalance{}, ...]
-
-  """
-  def list_bits_balances do
-    BitsBalanceQueries.all()
-    |> Repo.all()
-  end
-
-  @doc """
-  Gets a single bits_balance.
-
-  Raises `Ecto.NoResultsError` if the Bits balance does not exist.
-
-  ## Examples
-
-      iex> get_bits_balance!(123)
-      %BitsBalance{}
-
-      iex> get_bits_balance!(456)
-      ** (Ecto.NoResultsError)
-
-      iex> get_bits_balance!(user)
-      %BitsBalance{}
-
-  """
-  def get_bits_balance!(id) when is_integer(id) do
-    BitsBalanceQueries.with_id(id)
-    |> limit(1)
-    |> Repo.one!()
-  end
-
-  def get_bits_balance!(%User{} = user) do
-    BitsBalanceQueries.with_user_id(user.id)
-    |> limit(1)
-    |> Repo.one!()
-  end
-
-  @decorate cacheable(
-              cache: Cache,
-              key: {BitsBalance, user.id},
-              opts: [ttl: @cache_ttl]
-            )
-  def get_bits_balance_for_user(%User{} = user) do
-    BitsBalanceQueries.with_user_id(user.id)
-    |> limit(1)
-    |> Repo.one()
-  end
-
-  @doc """
-  Gets a single bits_balance by user_id.
-
-  ## Examples
-
-      iex> get_bits_balance_by_user_id(123)
-      {:ok, %BitsBalance{}
-
-      iex> get_bits_balance_by_user_id(456)
-      {:error, nil}
-  """
-  def get_bits_balance_by_user_id(user_id) do
-    BitsBalanceQueries.with_user_id(user_id)
-    |> limit(1)
-    |> Repo.one()
-    |> case do
-      nil -> {:error, nil}
-      bits_balance -> {:ok, bits_balance}
-    end
-  end
-
-  @doc """
-  Creates a bits_balance.
-
-  ## Examples
-
-      iex> create_bits_balance(user, %{})
-      {:ok, %BitsBalance{}}
-
-      iex> create_bits_balance(user, %{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def create_bits_balance(%User{} = user, attrs \\ %{}) do
-    user
-    |> Ecto.build_assoc(:bits_balance)
-    |> BitsBalance.changeset(attrs)
-    |> Repo.insert()
-  end
-
-  @doc """
-  Updates a bits_balance.
-
-  ## Examples
-
-      iex> update_bits_balance(bits_balance, %{field: new_value})
-      {:ok, %BitsBalance{}}
-
-      iex> update_bits_balance(bits_balance, %{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  @decorate cache_evict(
-              cache: Cache,
-              key: {BitsBalance, bits_balance.user_id}
-            )
-  @decorate cache_evict(
-              cache: Cache,
-              key: {BitsBalanceDebit, bits_balance.user_id}
-            )
-  def update_bits_balance(%BitsBalance{} = bits_balance, attrs) do
-    bits_balance
-    |> BitsBalance.update_changeset(attrs)
-    |> Repo.update()
-  end
-
-  @doc """
-  Deletes a bits_balance.
-
-  ## Examples
-
-      iex> delete_bits_balance(bits_balance)
-      {:ok, %BitsBalance{}}
-
-      iex> delete_bits_balance(bits_balance)
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def delete_bits_balance(%BitsBalance{} = bits_balance) do
-    Repo.delete(bits_balance)
-  end
-
-  @doc """
-  Returns an `%Ecto.Changeset{}` for tracking bits_balance changes.
-
-  ## Examples
-
-      iex> change_bits_balance(bits_balance)
-      %Ecto.Changeset{data: %BitsBalance{}}
-
-  """
-  def change_bits_balance(%BitsBalance{} = bits_balance, attrs \\ %{}) do
-    BitsBalance.changeset(bits_balance, attrs)
-  end
-
-  alias StreamClosedCaptionerPhoenix.Bits.BitsTransaction
-
-  @doc """
-  Returns the list of bits_transactions.
-
-  ## Examples
-
-      iex> list_bits_transactions()
-      [%BitsTransaction{}, ...]
-
-  """
-  def list_bits_transactions do
-    BitsTransactionQueries.all()
-    |> Repo.all()
-  end
-
-  @doc """
-  Gets a single bits_transaction.
-
-  Raises `Ecto.NoResultsError` if the Bits transactions does not exist.
-
-  ## Examples
-
-      iex> get_bits_transaction!(123)
-      %BitsTransaction{}
-
-      iex> get_bits_transaction!(456)
-      ** (Ecto.NoResultsError)
-
-  """
-  def get_bits_transaction!(id) when is_integer(id) do
-    BitsTransactionQueries.with_id(id)
-    |> limit(1)
-    |> Repo.one!()
-  end
-
-  def get_bits_transactions!(%User{} = user) do
-    BitsTransactionQueries.with_user_id(user.id)
-    |> Repo.all()
-  end
-
-  @doc """
-  Gets a single bits_transaction by its transaction_id.
-
-  ## Examples
-
-      iex> get_bits_transaction_by("123")
-      %BitsTransaction{}
-
-      iex> get_bits_transaction_by("456")
-      nil
-
-  """
-  def get_bits_transaction_by(transaction_id) when is_binary(transaction_id) do
-    BitsTransactionQueries.with_transaction_id(transaction_id)
-    |> limit(1)
-    |> Repo.one()
-  end
-
-  @doc """
-  Creates a bits_transaction.
-
-  ## Examples
-
-      iex> create_bits_transaction(user, %{field: value})
-      {:ok, %BitsTransaction{}}
-
-      iex> create_bits_transaction(user, %{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def create_bits_transaction(user, attrs \\ %{}) do
-    user
-    |> Ecto.build_assoc(:bits_transactions)
-    |> BitsTransaction.changeset(attrs)
-    |> Repo.insert()
-  end
-
-  @doc """
-  Deletes a bits_transaction.
-
-  ## Examples
-
-      iex> delete_bits_transaction(bits_transaction)
-      {:ok, %BitsTransaction{}}
-
-      iex> delete_bits_transaction(bits_transaction)
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def delete_bits_transaction(%BitsTransaction{} = bits_transaction) do
-    Repo.delete(bits_transaction)
-  end
-
-  @doc """
-  Returns an `%Ecto.Changeset{}` for tracking bits_transaction changes.
-
-  ## Examples
-
-      iex> change_bits_transaction(bits_transaction)
-      %Ecto.Changeset{data: %BitsTransaction{}}
-
-  """
-  def change_bits_transaction(%BitsTransaction{} = bits_transaction, attrs \\ %{}) do
-    BitsTransaction.changeset(bits_transaction, attrs)
-  end
-
-  @doc """
-  Procceses a transaction from Twitch that will credit a users bits balance account
-  and record the transaction.
-
-  ## Examples
-
-      iex> process_bits_transaction(uid, transaction_data)
-      %{:ok, transaction_multi_map}
-
-      iex> process_bits_transaction(uid, transaction_data)
-      %{:error, :validate_transaction, "Transaction 1 is already recorded.", %{}}
-  """
-  def process_bits_transaction(uid, decoded_token) do
-    transaction_info = decoded_token |> Map.get("data")
-    transaction_id = Map.get(transaction_info, "transactionId")
-    amount = get_in(transaction_info, ["product", "cost", "amount"])
-
-    result =
-      Ecto.Multi.new()
-      |> Ecto.Multi.run(:validate_amount, fn _repo, _ ->
-        if amount == 500 do
-          {:ok, amount}
-        else
-          {:error, :invalid_amount}
-        end
-      end)
-      |> Ecto.Multi.run(:validate_transaction, validate_transaction(transaction_id))
-      |> Ecto.Multi.run(:retrieve_channel_user, retrieve_user_by_uid(uid))
-      |> Ecto.Multi.run(:retrieve_balance, &retrieve_balance/2)
-      |> Ecto.Multi.run(:add_to_balance, add_to_balance(amount))
-      |> Ecto.Multi.run(:save_transaction, save_transaction(transaction_info))
-      |> Ecto.Multi.run(:publish_activity, &publish_activity/2)
-      |> Repo.transaction()
-
-    case result do
-      {:ok, %{retrieve_channel_user: user, add_to_balance: bits_balance}} ->
-        AuditLog.info("bits.credit_applied", %{
-          user_id: user.id,
-          transaction_id: transaction_id,
-          amount: amount,
-          balance: bits_balance.balance
-        })
-
-      {:error, step, reason, _} ->
-        AuditLog.warn("bits.credit_apply_failed", %{
-          uid: uid,
-          transaction_id: transaction_id,
-          step: step,
-          reason: audit_failure_reason(reason)
-        })
-    end
-
-    result
-  end
-
-  defp validate_transaction(transaction_id) do
-    fn _repo, _ ->
-      case get_bits_transaction_by(transaction_id) do
-        nil -> {:ok, transaction_id}
-        _transaction -> {:error, "Transaction #{transaction_id} is already recorded."}
-      end
-    end
-  end
-
-  defp retrieve_user_by_uid(uid) do
-    fn _repo, _ ->
-      case StreamClosedCaptionerPhoenix.AccountsOauth.get_user_for_provider("twitch", uid) do
-        nil ->
-          {:error, "Channel #{uid} not found"}
-
-        user ->
-          {:ok, user}
-      end
-    end
-  end
-
-  defp retrieve_balance(_repo, %{retrieve_channel_user: user}) do
-    user = Repo.preload(user, :bits_balance)
-
-    if is_nil(user.bits_balance) do
-      {:error, :no_bits_balance}
-    else
-      {:ok, user.bits_balance}
-    end
-  end
-
-  defp add_to_balance(amount) do
-    fn _repo, %{retrieve_balance: bits_balance} ->
-      update_bits_balance(bits_balance, %{balance: bits_balance.balance + amount})
-    end
-  end
-
-  defp save_transaction(transaction_info) do
-    fn _repo, %{retrieve_channel_user: user} ->
-      create_bits_transaction(user, %{
-        amount: get_in(transaction_info, ["product", "cost", "amount"]),
-        purchaser_uid: get_in(transaction_info, ["userId"]),
-        sku: get_in(transaction_info, ["product", "sku"]),
-        # get_in(transaction_info, ["time"]),
-        time: Timex.now(),
-        transaction_id: get_in(transaction_info, ["transactionId"])
-      })
-    end
-  end
-
-  defp audit_failure_reason(reason) when is_atom(reason), do: reason
-  defp audit_failure_reason(%Ecto.Changeset{} = cs), do: Ecto.Changeset.traverse_errors(cs, fn {msg, _} -> msg end)
-  defp audit_failure_reason(reason) when is_binary(reason), do: reason
-  defp audit_failure_reason(reason), do: inspect(reason, limit: 5, printable_limit: 100)
-
-  defp publish_activity(_repo, %{retrieve_channel_user: user, add_to_balance: bits_balance}) do
-    StreamClosedCaptionerPhoenixWeb.Endpoint.broadcast(
-      "captions:#{user.id}",
-      "transaction",
-      %{balance: bits_balance.balance}
-    )
-
-    {:ok, nil}
-  end
+  @moduledoc "The Bits context — thin facade delegating to focused sub-contexts."
+
+  alias StreamClosedCaptionerPhoenix.Bits.Balance
+  alias StreamClosedCaptionerPhoenix.Bits.Debit
+  alias StreamClosedCaptionerPhoenix.Bits.Transaction
+
+  # Balance
+  defdelegate list_bits_balances(), to: Balance
+  defdelegate get_bits_balance!(id), to: Balance
+  defdelegate get_bits_balance_for_user(user), to: Balance
+  defdelegate get_bits_balance_by_user_id(user_id), to: Balance
+  defdelegate create_bits_balance(user), to: Balance
+  defdelegate create_bits_balance(user, attrs), to: Balance
+  defdelegate update_bits_balance(bits_balance, attrs), to: Balance
+  defdelegate delete_bits_balance(bits_balance), to: Balance
+  defdelegate change_bits_balance(bits_balance), to: Balance
+  defdelegate change_bits_balance(bits_balance, attrs), to: Balance
+
+  # Debit
+  defdelegate list_bits_balance_debits(), to: Debit
+  defdelegate list_users_bits_balance_debits(user), to: Debit
+  defdelegate get_bits_balance_debit!(id), to: Debit
+  defdelegate get_users_bits_balance_debit!(user, id), to: Debit
+  defdelegate get_user_active_debit(user_id), to: Debit
+  defdelegate get_translation_snapshot(user_id), to: Debit
+  defdelegate user_active_debit_exists?(user_id), to: Debit
+  defdelegate create_bits_balance_debit(user), to: Debit
+  defdelegate create_bits_balance_debit(user, attrs), to: Debit
+  defdelegate activate_translations_for(user), to: Debit
+
+  # Transaction
+  defdelegate bits_transactions_and_debits_for_user(user_id, offset, limit), to: Transaction
+  defdelegate list_bits_transactions(), to: Transaction
+  defdelegate get_bits_transaction!(id), to: Transaction
+  defdelegate get_bits_transactions!(user), to: Transaction
+  defdelegate get_bits_transaction_by(transaction_id), to: Transaction
+  defdelegate create_bits_transaction(user), to: Transaction
+  defdelegate create_bits_transaction(user, attrs), to: Transaction
+  defdelegate delete_bits_transaction(bits_transaction), to: Transaction
+  defdelegate change_bits_transaction(bits_transaction), to: Transaction
+  defdelegate change_bits_transaction(bits_transaction, attrs), to: Transaction
+  defdelegate process_bits_transaction(uid, decoded_token), to: Transaction
 end
